# src/MedicalOrders.Api/Contracts/CreateOrderRequest.cs
namespace MedicalOrders.Api.Contracts;

public sealed record CreateOrderRequest(
    string? PatientId,
    string? PatientName,
    string? ServiceCode,
    string? ServiceDescription,
    string? Priority);